+++ google/resource_compute_disk_test.go
@@ -12,6 +12,150 @@ import (
 	"google.golang.org/api/compute/v1"
 )
 
+func TestDiskImageDiffSuppress(t *testing.T) {
+	cases := map[string]struct {
+		Old, New           string
+		ExpectDiffSuppress bool
+	}{
+		// Full & partial links
+		"matching self_link with different api version": {
+			Old:                "https://www.googleapis.com/compute/beta/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			ExpectDiffSuppress: true,
+		},
+		"matching image partial self_link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			ExpectDiffSuppress: true,
+		},
+		"matching image partial no project self_link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "global/images/debian-8-jessie-v20171213",
+			ExpectDiffSuppress: true,
+		},
+		"different image self_link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-7-jessie-v20171213",
+			ExpectDiffSuppress: false,
+		},
+		"different image partial self_link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "projects/debian-cloud/global/images/debian-7-jessie-v20171213",
+			ExpectDiffSuppress: false,
+		},
+		"different image partial no project self_link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "global/images/debian-7-jessie-v20171213",
+			ExpectDiffSuppress: false,
+		},
+		// Image name
+		"matching image name": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "debian-8-jessie-v20171213",
+			ExpectDiffSuppress: true,
+		},
+		"different image name": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "debian-7-jessie-v20171213",
+			ExpectDiffSuppress: false,
+		},
+		// Image short hand
+		"matching image short hand": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "debian-cloud/debian-8-jessie-v20171213",
+			ExpectDiffSuppress: true,
+		},
+		"matching image short hand but different project": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "different-cloud/debian-8-jessie-v20171213",
+			ExpectDiffSuppress: false,
+		},
+		"different image short hand": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "debian-cloud/debian-7-jessie-v20171213",
+			ExpectDiffSuppress: false,
+		},
+		// Latest image short hand
+		"matching latest image short hand": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "debian-cloud/debian-8",
+			ExpectDiffSuppress: true,
+		},
+		"matching latest image short hand with project short name": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "debian/debian-8",
+			ExpectDiffSuppress: true,
+		},
+		"matching latest image short hand but different project": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "different-cloud/debian-8",
+			ExpectDiffSuppress: false,
+		},
+		"different latest image short hand": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "debian-cloud/debian-7",
+			ExpectDiffSuppress: false,
+		},
+		// Image Family
+		"matching image family": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "family/debian-8",
+			ExpectDiffSuppress: true,
+		},
+		"matching image family self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/family/debian-8",
+			ExpectDiffSuppress: true,
+		},
+		"matching image family partial self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "projects/debian-cloud/global/images/family/debian-8",
+			ExpectDiffSuppress: true,
+		},
+		"matching image family partial no project self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "global/images/family/debian-8",
+			ExpectDiffSuppress: true,
+		},
+		"different image family": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "family/debian-7",
+			ExpectDiffSuppress: false,
+		},
+		"different image family self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/family/debian-7",
+			ExpectDiffSuppress: false,
+		},
+		"different image family partial self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "projects/debian-cloud/global/images/family/debian-7",
+			ExpectDiffSuppress: false,
+		},
+		"different image family partial no project self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "global/images/family/debian-7",
+			ExpectDiffSuppress: false,
+		},
+		"matching image family but different project in self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "https://www.googleapis.com/compute/v1/projects/other-cloud/global/images/family/debian-8",
+			ExpectDiffSuppress: false,
+		},
+		"different image family but different project in partial self link": {
+			Old:                "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20171213",
+			New:                "projects/other-cloud/global/images/family/debian-8",
+			ExpectDiffSuppress: false,
+		},
+	}
+
+	for tn, tc := range cases {
+		if diskImageDiffSuppress("image", tc.Old, tc.New, nil) != tc.ExpectDiffSuppress {
+			t.Errorf("bad: %s, %q => %q expect DiffSuppress to return %t", tn, tc.Old, tc.New, tc.ExpectDiffSuppress)
+		}
+	}
+}
+
 func TestAccComputeDisk_basic(t *testing.T) {
 	t.Parallel()
 
@@ -31,6 +175,11 @@ func TestAccComputeDisk_basic(t *testing.T) {
 					testAccCheckComputeDiskHasLabel(&disk, "my-label", "my-label-value"),
 					testAccCheckComputeDiskHasLabelFingerprint(&disk, "google_compute_disk.foobar"),
 				),
+			},
+			resource.TestStep{
+				ResourceName:      "google_compute_disk.foobar",
+				ImportState:       true,
+				ImportStateVerify: true,
 			},
 		},
 	})