+++ google/resource_compute_health_check_test.go
@@ -33,6 +33,11 @@ func TestAccComputeHealthCheck_tcp(t *testing.T) {
 					testAccCheckComputeHealthCheckTcpPort(80, &healthCheck),
 				),
 			},
+			resource.TestStep{
+				ResourceName:      "google_compute_health_check.foobar",
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }
@@ -94,6 +99,11 @@ func TestAccComputeHealthCheck_ssl(t *testing.T) {
 						3, 3, &healthCheck),
 				),
 			},
+			resource.TestStep{
+				ResourceName:      "google_compute_health_check.foobar",
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }
@@ -119,6 +129,11 @@ func TestAccComputeHealthCheck_http(t *testing.T) {
 						3, 3, &healthCheck),
 				),
 			},
+			resource.TestStep{
+				ResourceName:      "google_compute_health_check.foobar",
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }
@@ -143,6 +158,11 @@ func TestAccComputeHealthCheck_https(t *testing.T) {
 					testAccCheckComputeHealthCheckThresholds(
 						3, 3, &healthCheck),
 				),
+			},
+			resource.TestStep{
+				ResourceName:      "google_compute_health_check.foobar",
+				ImportState:       true,
+				ImportStateVerify: true,
 			},
 		},
 	})