+++ google/resource_compute_subnetwork_test.go
@@ -35,6 +35,16 @@ func TestAccComputeSubnetwork_basic(t *testing.T) {
 						"google_compute_subnetwork.network-ref-by-name", &subnetwork2),
 				),
 			},
+			resource.TestStep{
+				ResourceName:      "google_compute_subnetwork.network-ref-by-url",
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+			resource.TestStep{
+				ResourceName:      "google_compute_subnetwork.network-with-private-google-access",
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }
@@ -74,7 +84,7 @@ func TestAccComputeSubnetwork_update(t *testing.T) {
 	}
 }
 
-func TestAccComputeSubnetwork_secondaryIpRangesUpdate(t *testing.T) {
+func TestAccComputeSubnetwork_secondaryIpRanges(t *testing.T) {
 	t.Parallel()
 
 	var subnetwork compute.Subnetwork
@@ -103,35 +113,11 @@ func TestAccComputeSubnetwork_secondaryIpRangesUpdate(t *testing.T) {
 				),
 			},
 			resource.TestStep{
-				Config: testAccComputeSubnetwork_secondaryIpRanges_update3(cnName, subnetworkName),
+				Config: testAccComputeSubnetwork_secondaryIpRanges_update1(cnName, subnetworkName),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckComputeSubnetworkExists("google_compute_subnetwork.network-with-private-secondary-ip-ranges", &subnetwork),
 					testAccCheckComputeSubnetworkHasSecondaryIpRange(&subnetwork, "tf-test-secondary-range-update1", "192.168.10.0/24"),
 					testAccCheckComputeSubnetworkHasNotSecondaryIpRange(&subnetwork, "tf-test-secondary-range-update2", "192.168.11.0/24"),
-				),
-			},
-		},
-	})
-}
-
-func TestAccComputeSubnetwork_secondaryIpRanges(t *testing.T) {
-	t.Parallel()
-
-	var subnetwork compute.Subnetwork
-
-	cnName := fmt.Sprintf("tf-test-%s", acctest.RandString(10))
-	subnetworkName := fmt.Sprintf("tf-test-%s", acctest.RandString(10))
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckComputeSubnetworkDestroy,
-		Steps: []resource.TestStep{
-			resource.TestStep{
-				Config: testAccComputeSubnetwork_secondaryIpRanges(cnName, subnetworkName),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckComputeSubnetworkExists("google_compute_subnetwork.network-with-private-secondary-ip-range", &subnetwork),
-					testAccCheckComputeSubnetworkHasSecondaryIpRange(&subnetwork, "tf-test-secondary-range", "192.168.1.0/24"),
 				),
 			},
 		},
@@ -281,26 +267,6 @@ resource "google_compute_subnetwork" "network-with-private-google-access" {
 `, cnName, subnetworkName)
 }
 
-func testAccComputeSubnetwork_secondaryIpRanges(cnName, subnetworkName string) string {
-	return fmt.Sprintf(`
-resource "google_compute_network" "custom-test" {
-	name = "%s"
-	auto_create_subnetworks = false
-}
-
-resource "google_compute_subnetwork" "network-with-private-secondary-ip-range" {
-	name = "%s"
-	ip_cidr_range = "10.2.0.0/16"
-	region = "us-central1"
-	network = "${google_compute_network.custom-test.self_link}"
-	secondary_ip_range {
-		range_name = "tf-test-secondary-range"
-		ip_cidr_range = "192.168.1.0/24"
-	}
-}
-`, cnName, subnetworkName)
-}
-
 func testAccComputeSubnetwork_secondaryIpRanges_update1(cnName, subnetworkName string) string {
 	return fmt.Sprintf(`
 resource "google_compute_network" "custom-test" {
@@ -341,26 +307,6 @@ resource "google_compute_subnetwork" "network-with-private-secondary-ip-ranges" 
 		range_name = "tf-test-secondary-range-update2"
 		ip_cidr_range = "192.168.11.0/24"
 	},
-}
-`, cnName, subnetworkName)
-}
-
-func testAccComputeSubnetwork_secondaryIpRanges_update3(cnName, subnetworkName string) string {
-	return fmt.Sprintf(`
-resource "google_compute_network" "custom-test" {
-	name = "%s"
-	auto_create_subnetworks = false
-}
-
-resource "google_compute_subnetwork" "network-with-private-secondary-ip-ranges" {
-	name = "%s"
-	ip_cidr_range = "10.2.0.0/16"
-	region = "us-central1"
-	network = "${google_compute_network.custom-test.self_link}"
-	secondary_ip_range {
-		range_name = "tf-test-secondary-range-update1"
-		ip_cidr_range = "192.168.10.0/24"
-	}
 }
 `, cnName, subnetworkName)
 }